requirements.txt
pandas
boto3
pyarrow
prefect
psycopg2-binary